+++ winim/ole.nim
@@ -9,5 +9,5 @@
 
 {.deadCodeElim: on.}
 
-import inc/[objbase, objext, uiautomation]
-export objbase, objext, uiautomation
+import inc/[objbase, objext, uiautomation, mscoree]
+export objbase, objext, uiautomation, mscoree